fix bid price rounding, organize bid tests
